Add an example project for "unit tests"
This adds an example project that makes it a little easier to
contribute to development of react-native-blob-util.

The idea is that you can write unit tests that run on a simulator or
real device, providing a simple way to check if functionality works as
expected. It runs the tests sequentially when you press the "Run tests"
button in the app and displays the results after.

While not as efficient as using JUnit or similar for unit testing, it's
more realistic as it runs on a simulator/device and it's perhaps easier
for React Native developers to contribute to, as JavaScript/React is
probably more familiar for most.

I'm not that familiar with Jest or React Testing Library, but I guess
there could be some way of using these testing tools in order to run
tests more automatically, or on CI.

I have added some tests that illustrate some bugs in this library. For
instance the issue
[#180](https://github.com/RonRadtke/react-native-blob-util/issues/180)
is reproduced.